requirements.txt
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.1
paramiko==2.9.2
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.19.2
scp==0.14.2
six==1.16.0
